Created an Operation base class, added two operation types
