+++ src/com/company/banking/models/accounts/RegularAccount.java
@@ -1,4 +1,4 @@
-package com.company.banking.models.Account;
+package com.company.banking.models.accounts;
 
 public class RegularAccount extends Account {
     public RegularAccount() {}
