+++ src/com/company/banking/models/operations/Transaction.java
@@ -1,26 +1,14 @@
-package com.company.banking.models;
+package com.company.banking.models.operations;
 
-import com.company.banking.models.Account.Account;
+import com.company.banking.models.accounts.Account;
 
-import java.util.UUID;
-
-public class Transaction {
-    private UUID id;
-    private final double amount;
+public class Transaction extends Operation {
     private final Account payerAccount, payeeAccount;
 
     public Transaction(double amount, Account payerAccount, Account payeeAccount) {
-        this.amount = amount;
+        super(amount);
         this.payerAccount = payerAccount;
         this.payeeAccount = payeeAccount;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public double getAmount() {
-        return amount;
     }
 
     public Account getPayerAccount() {
@@ -29,10 +17,6 @@ public class Transaction {
 
     public Account getPayeeAccount() {
         return payeeAccount;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
     }
 
     @Override
